Bug(s) fixed: 186280 Bug Description: ldapserver: Close potential security vulnerabilities in CGI code Reviewed by: Nathan, Noriko, and Pete (Thanks!) Fix Description: Clean up usage of sprintf, strcpy, fgets instead of gets, fixed buffer usage, etc., mostly in the CGI code and other user facing code (i.e. setup).  Also, Steve Grubb told me about a GCC trick to force it to check printf style varargs functions, to check the format string against the argument string, for type mismatches, missing arguments, and too many arguments. In the CGI form argument parsing code, we needed to be more careful about checking for bad input - good input is supposed to look like this: name=value&name=value&..... &name=value.  I don't think the original code was checking properly for something like name&name=value. There was another place where we were not checking to see if a buffer had enough room before appending a string to it. I had to change a couple of functions to allow passing in the size of the buffer. Fixed some issues raised by Noriko and Nathan. Platforms tested: RHEL4 Flag Day: no Doc impact: no QA impact: should be covered by regular nightly and manual testing New Tests integrated into TET: none

+++ ldap/servers/plugins/chainingdb/cb_monitor.c
@@ -76,9 +76,9 @@ cb_search_monitor_callback(Slapi_PBlock * pb, Slapi_Entry * e, Slapi_Entry * ent
 	char          		buf[CB_BUFSIZE];
 	struct berval 		val;
   	struct berval 		*vals[2];
-	int 			deletecount,addcount,modifycount,modrdncount,searchbasecount,searchonelevelcount;
-	int 			searchsubtreecount,abandoncount,bindcount,unbindcount,comparecount;
-	int 			outgoingconn, outgoingbindconn;
+	unsigned long		deletecount,addcount,modifycount,modrdncount,searchbasecount,searchonelevelcount;
+	unsigned long		searchsubtreecount,abandoncount,bindcount,unbindcount,comparecount;
+	unsigned int 		outgoingconn, outgoingbindconn;
 	cb_backend_instance	*inst = (cb_backend_instance *)arg;
 
 	/* First make sure the backend instance is configured */
@@ -178,12 +178,12 @@ cb_search_monitor_callback(Slapi_PBlock * pb, Slapi_Entry * e, Slapi_Entry * ent
   	val.bv_len = strlen( buf );
   	slapi_entry_attr_replace( e, CB_MONITOR_COMPARECOUNT, ( struct berval **)vals );
 
-  	sprintf( buf, "%d", outgoingconn );
+  	sprintf( buf, "%u", outgoingconn );
   	val.bv_val = buf;
   	val.bv_len = strlen( buf );
   	slapi_entry_attr_replace( e, CB_MONITOR_OUTGOINGCONN, ( struct berval **)vals );
 
-  	sprintf( buf, "%d", outgoingbindconn );
+  	sprintf( buf, "%u", outgoingbindconn );
   	val.bv_val = buf;
   	val.bv_len = strlen( buf );
   	slapi_entry_attr_replace( e, CB_MONITOR_OUTGOINGBINDCOUNT, ( struct berval **)vals );
